only use jq docker image if not installed locally
Former-commit-id: de9e717c9d69b32ea4e91eb0a6841d40517474f7 [formerly a8c104dd6825624b0daa76650d0af2a0956357b3 [formerly 642971f56211f97e364873fcc9c4ac77d7ed624d]]
Former-commit-id: f7525862524c156b6bca07a5fed6a75d1a917a50
Former-commit-id: 7bd16fb027b829049209d26cb9b6e26201361893

+++ scripts/utils.sh
@@ -490,7 +490,12 @@ function timeout() {
 }
 
 function jq() {
-    docker run --rm -i imega/jq "$@"
+    if [[ $(type -f jq 2>&1) =~ "not found" ]]
+    then
+      docker run --rm -i imega/jq "$@"
+    else
+      $(which jq) "$@"
+    fi
 }
 
 function az() {
